Extract validator reference helpers
